Fixed many compiler errors and some warnings.
Compiles and runs on:
Clang 4.0
GCC 4.7
GCC 7.0.1
VS 2015

+++ include/Ace/Sprite.h
@@ -11,7 +11,7 @@ namespace ace
     {
     public:
 
-		static const UInt32 size = 4;
+		static const UInt8 size = 4u;
 
 		std::array<Vertex, size> vertexData;
